// code/frontend/CNet.Web.Admin/src/api/admin/gen/data-contracts.ts
export type ResCode = 0 | 1 | -1
// ...
export interface BooleanDataRes {
  code?: ResCode
  msg?: string | null
  data?: boolean
}
// ...
export interface PageDataReq {
  /** @format int32 */
  pageNum?: number
  /** @format int32 */
  pageSize?: number
  field?: string | null
  order?: string | null
  query?: Record<string, any>
}

export interface PubRolePageDateRes {
  code?: ResCode
  msg?: string | null
  /** @format int32 */
  count?: number
  /** @format int32 */
  totalPage?: number
  data?: PubRole[] | null
  /** @format int32 */
  pageNum?: number
  /** @format int32 */
  pageSize?: number
}

export interface StringIEnumerableDataRes {
  code?: ResCode
  msg?: string | null
  data?: string[] | null
}

export interface PubRoleDataRes {
  code?: ResCode
  msg?: string | null
  data?: PubRole
}

export interface VPubuserDept {
  roleCodes?: string[] | null
  roleNames?: string[] | null
  /** @format int32 */
  id?: number
  userCode?: string | null
  userName?: string | null
  realName?: string | null
  userPwd?: string | null
  sex?: boolean
  identityNo?: string | null
  /** @format date-time */
  birthday?: string | null
  deptCode?: string | null
  managerFlag?: boolean
  tel?: string | null
  eMail?: string | null
  qq?: string | null
  remark?: string | null
  stopFlag?: boolean
  crid?: string | null
  /** @format date-time */
  crdt?: string | null
  lmid?: string | null
  /** @format date-time */
  lmdt?: string | null
  /** @format date-time */
  loginDate?: string | null
  provinceCode?: string | null
  cityCode?: string | null
  regionCode?: string | null
  userAddress?: string | null
  wxcode?: string | null
  headUrl?: string | null
  /** @format int32 */
  userStatus?: number | null
  deptName?: string | null
}

export interface VPubuserDeptPageDateRes {
  code?: ResCode
  msg?: string | null
  /** @format int32 */
  count?: number
  /** @format int32 */
  totalPage?: number
  data?: VPubuserDept[] | null
  /** @format int32 */
  pageNum?: number
  /** @format int32 */
  pageSize?: number
}

export interface VPubuserDeptDataRes {
  code?: ResCode
  msg?: string | null
  data?: VPubuserDept
}